feat: add track activity management
- Updated the service interface to include methods for searching tracks with user ID and upserting track activities.
- Implemented the UpsertTrackActivities handler to manage user track activities.
- Enhanced the Search functionality to include user-specific track activity data.
- Created a new TrackActivity model and repository for managing track activities in the database.
- Added tests for the new track activities functionality in both service and handler layers.
- Updated JWT token creation to use uint for user IDs instead of int64.
- Modified Spotify track model to include an IsLiked field for user preferences.

+++ internal/service/memberships/login.go
@@ -26,7 +26,7 @@ func (s *service) Login(request memberships.LoginRequest) (string, error) {
 		return "", errors.New("email and password not match")
 	}
 
-	accessToken, err := jwt.CreateToken(int64(userDetail.ID), userDetail.Username, s.cfg.Service.SecretJWT)
+	accessToken, err := jwt.CreateToken(userDetail.ID, userDetail.Username, s.cfg.Service.SecretJWT)
 	if err != nil {
 		log.Error().Err(err).Msg("failed to create token")
 		return "", err
